# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AidFlow API",
    "description": "AidFlow — 災害情境驅動的災情通報、案件處理、派工與視覺化平台生成器。公開端點位於 /v1/public；其餘為政府後台端點。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/demo/nantou": {
      "post": {
        "tags": [
          "demo"
        ],
        "summary": "Seed the 南投豪雨 demo scenario through the real pipeline",
        "operationId": "seed_nantou_v1_demo_nantou_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Seed Nantou V1 Demo Nantou Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Locate the demo platform",
        "operationId": "get_demo_v1_demo_nantou_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Demo V1 Demo Nantou Get"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "設定 ADMIN_API_KEY 後，/v1/public 與 /v1/health 以外的端點需帶此標頭。"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}